requirements.txt
pytz
slack_sdk
python-dotenv
